small updates

+++ calculater/arithmetic.h
@@ -1,7 +1,7 @@
 #ifndef __MAIN__H__
 #define __MAIN__H__
 int Add(int first_summand, int second_summand, int* result);
-double Div(int dividend, int divider, double* result);
+int Div(int dividend, int divider, double* result);
 int Mult(int first_multiplier, int second_multiplier, int* result);
 int Sub(int minuend, int subtrahend, int* result);
 #endif
